requirements.txt
#certifi          ==2018.11.29 
#chardet          ==3.0.4      
#cloudpickle      ==0.6.1      
#cycler           ==0.10.0     
#cytoolz          ==0.9.0.1    
#dask             ==1.0.0      
#decorator        ==4.3.0      
easydict         ==1.9        
#functools32      ==3.2.3.post2
#futures          ==3.2.0      
#graphviz         ==0.8.4      
#imageio          ==2.4.1      
#kiwisolver       ==1.0.1      
matplotlib       ==2.2.3      
mxnet            ==1.3.1      
#networkx         ==2.2        
#olefile          ==0.46       
opencv-python    ==3.4.5.20   
#Pillow           ==5.4.0      
#pip              ==18.1       
#pyparsing        ==2.3.0      
#python-dateutil  ==2.7.5      
#pytz             ==2018.7     
#PyWavelets       ==1.0.1      
#requests         ==2.21.0     
scikit-image     ==0.14.1     
scikit-learn     ==0.20.2     
scipy            ==1.1.0      
#setuptools       ==40.6.3     
#singledispatch   ==3.4.0.3    
#six              ==1.12.0     
#subprocess32     ==3.5.3      
#toolz            ==0.9.0      
#tornado          ==5.1.1      
#urllib3          ==1.24.1     
#wheel            ==0.32.3
protobuf          ==3.6.1
grpcio            ==1.17.1
grpcio-tools      ==1.17.1